Predict IBNR frequency

This function predicts the results from the ReSurv fits.

# S3 method for class 'ReSurvFit'
predict(
  object,
  newdata = NULL,
  grouping_method = "probability",
  check_value = 1.85,
  ...
)

Arguments

object

ResurvFit object specifying start time, end time and status.

newdata

IndividualDataPP object that contains new data to predict.

grouping_method

character, use probability or exposure approach to group from input to output development factors. Choice between:

"exposure"

"probability"

Default is "exposure".

check_value

numeric, check hazard value on initial granularity, if above threshold we increase granularity to try and adjust the development factor.

...

Additional arguments to pass to the predict function.

Value

Predictions for the ReSurvFit model. It includes

ReSurvFit: Fitted ReSurv model.

long_triangle_format_out: data.frame. Predicted development factors and IBNR claim counts for each feature combination in long format.

input_granularity: data.frame. Predictions for each feature combination in long format for input_time_granularity.

AP_i: Accident period, input_time_granularity.

DP_i: Development period, input_time_granularity.

f_i: Predicted development factors, input_time_granularity.

group_i: Group code, input_time_granularity. This associates to each feature combination an identifier.

expected_counts: Expected counts, input_time_granularity.

IBNR: Predicted IBNR claim counts, input_time_granularity.

output_granularity: data.frame. Predictions for each feature combination in long format for output_time_granularity.

AP_o: Accident period, output_time_granularity.

DP_o: Development period, output_time_granularity.

f_o: Predicted development factors, output_time_granularity.

group_o: Group code, output_time_granularity. This associates to each feature combination an identifier.

expected_counts: Expected counts, output_time_granularity.

IBNR: Predicted IBNR claim counts, output_time_granularity.

lower_triangle: Predicted lower triangle.

input_granularity: data.frame. Predicted lower triangle for input_time_granularity.

output_granularity: data.frame. Predicted lower triangle for output_time_granularity.

predicted_counts: numeric. Predicted total frequencies.

grouping_method: character. Chosen grouping method.